# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "node-koa-templates API",
    "version": "1.0.0",
    "summary": "test",
    "description": "token:   Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJBY2NvdW50SWQiOiI0ZTFiYjNmNS04YzBiLTQ1NzMtOGU2Zi1hYmY5M2JkYzQyMjQiLCJpYXQiOjE3MTAxNTA3MDEsImV4cCI6MTcxMjc0MjcwMX0.3_o-5nKLLnISV_fJ6n_-eAXX2r_S4W_04JFk3rL5RwU",
    "license": {
      "name": "MIT",
      "url": "https://github.com/saisilinus/node-express-mongoose-typescript-boilerplate.git"
    },
    "contact": {
      "name": "API Support",
      "url": "https://github.com/DuRuofu",
      "email": "duruofu@qq.com"
    }
  },
  "servers": [
    {
      "url": "http://172.31.32.1:3000/v1",
      "description": "API V1"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header"
      }
    }
  },
  "paths": {
...
    "/permission/permissions": {
      "post": {
        "summary": "添加一条权限",
        "description": "添加一条权限",
        "tags": [
          "权限管理"
        ],
        "security": [
          {
            "token": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Name": {
                    "description": "权限名称"
                  },
                  "Tag": {
                    "description": "权限标签"
                  },
                  "Type": {
                    "description": "权限类型"
                  },
                  "RuleValue": {
                    "description": "权限值"
                  },
                  "Description": {
                    "description": "描述"
                  },
                  "Action": {
                    "description": "行为"
                  }
                },
                "example": {
                  "Name": "添加用户",
                  "Tag": "用户管理",
                  "Type": "1",
                  "RuleValue": "/v1/user/users",
                  "Description": "添加用户",
                  "Action": "post"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "description": "状态码",
                      "example": 200
                    },
                    "massage": {
                      "type": "string",
                      "description": "注册信息",
                      "example": "用户注册成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "用户信息",
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/permission/permissions/{type}": {
      "get": {
        "summary": "获取系统权限列表",
        "description": "按类型获取权限列表",
        "tags": [
          "权限管理"
        ],
        "security": [
          {
            "token": {}
          }
        ],
        "parameters": [
          {
            "name": "type",
            "description": "类型【Type-->1:后端,2:前端】",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "description": "状态码",
                      "example": 200
                    },
                    "massage": {
                      "type": "string",
                      "description": "信息",
                      "example": "获取权限成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "权限信息",
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "重置系统权限列表",
        "description": "按类型获取权限列表",
        "tags": [
          "权限管理"
        ],
        "security": [
          {
            "token": {}
          }
        ],
        "parameters": [
          {
            "name": "type",
            "description": "类型【Type-->1:后端,2:前端】",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "number",
                      "description": "状态码",
                      "example": 200
                    },
                    "massage": {
                      "type": "string",
                      "description": "信息",
                      "example": "获取权限成功"
                    },
                    "data": {
                      "type": "object",
                      "description": "权限信息",
                      "example": null
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/permission/permissions/{id}": {
      "delete": {
        "summary": "删除某个权限",
        "description": "删除某个权限",
        "tags": [
          "权限管理"
        ],
        "security": [
          {
            "token": {}
          }
        ],
        "parameters": [
          {
            "name": "id",
            "description": "权限id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "number",
                  "description": "状态码",
                  "example": 200
                },
                "massage": {
                  "type": "string",
                  "description": "查询信息",
                  "example": "删除成功"
                },
                "data": {
                  "type": "object",
                  "description": "用户信息"
                }
              }
            }
          }
        }
      }
    },
    "/permission/permissions/roles/{id}": {
      "post": {
        "summary": "为用户添加角色",
        "description": "为用户添加角色",
        "tags": [
          "权限管理"
        ],
        "security": [
          {
            "token": {}
          }
        ],
        "parameters": [
          {
            "name": "id",
            "description": "权限id",
            "in": "path",
            "required": true
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "Roles": {
                    "description": "角色值数组"
                  }
                },
                "example": {
                  "Roles": [
                    "admin",
                    "admin1",
                    "admin2"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "number",
                  "description": "状态码",
                  "example": 200
                },
                "massage": {
                  "type": "string",
                  "description": "查询信息",
                  "example": "删除成功"
                },
                "data": {
                  "type": "object",
                  "description": "用户信息"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除用户所有角色",
        "description": "删除用户所有角色",
        "tags": [
          "权限管理"
        ],
        "security": [
          {
            "token": {}
          }
        ],
        "parameters": [
          {
            "name": "id",
            "description": "权限id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "number",
                  "description": "状态码",
                  "example": 200
                },
                "massage": {
                  "type": "string",
                  "description": "查询信息",
                  "example": "删除成功"
                },
                "data": {
                  "type": "object",
                  "description": "用户信息"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}